requirements.txt
requests
lxml
Pillow
mysql-connector-python
